separate MemberAvatar.js & MemberNickname.js
