--- public/css/Header/MemberNickname.scss
@@ -1,25 +1,0 @@
-.member-nickname {
-  float: left;
-  display: inline-block;
-  max-width: 220px;
-  color: $heading;
-  vertical-align: middle;
-  margin-right: 5px;
-  height: 40px;
-  line-height: 42px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.member-nickname > span {
-  transition: color .3s;
-  color: $font;
-}
-
-
-
-  @media (max-width: 768px) {
-    .member-nickname {max-width: 120px;}
-    .member-nickname > span { display: none; }
-  }
