+++ public/css/Header/MemberPanel.scss
@@ -7,8 +7,24 @@
   @extend %animated;
   @extend  %fadeInDown;
 
+
+  > .member-nickname {
+    float: left;
+    max-width: 220px;
+    margin-right: 5px;
+    height: 40px;
+    line-height: 42px;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+
+  .hi {
+    display: inline-block;
+  }
+
   &:hover {
-    > .member-nickname > span {
+    .hi {
       color: $active1;
     }
 
@@ -20,4 +36,18 @@
 
 .login  .member-panel {
   display: block;
+}
+
+
+@media (max-width: 768px) {
+  .member-panel  {
+    > .member-nickname {
+        max-width: 120px;
+      }
+
+      .hi {
+        display: none;
+      }
+
+    }
 }
