+++ public/css/LandingPart/SignForm.scss
@@ -7,7 +7,7 @@
   left: 0;
   margin: auto;
   width: 90%;
-  height: 80vh;
+  height: 80%;
   max-width: 400px;
   max-height: 480px;
   // animation-duration: .8s;
