+++ public/css/Shared/MemberAvatar.scss
