+++ public/css/Shared/MemberNickname.scss
@@ -1,0 +1,11 @@
+.member-nickname {
+  display: inline-block;
+  color: $heading;
+  vertical-align: middle;
+}
+
+.member-nickname > .hi {
+  display: none;
+  transition: color .3s;
+  color: $font;
+}
