+++ public/js/components/Shared/MemberAvatar.js
