+++ public/js/components/Shared/MemberNickname.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 
 const MemberNickname = () =>
   <span className="member-nickname">
-    <span>Hi,&nbsp;</span>
+    <span className='hi'>Hi,&nbsp;</span>
       Don Corleone
   </span>
